requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
azure-ai-inference==1.0.0b4
azure-core==1.29.5
supabase==2.3.4
python-dotenv==1.0.0
Pillow==10.1.0
